Feat: Model loading

+++ src/simulator/vulkan/indexBuffer.cpp
@@ -6,7 +6,7 @@
 //--------------------------------------------------
 #include "indexBuffer.h"
 
-IndexBuffer::IndexBuffer(Device* device, std::vector<uint16_t> indices):
+IndexBuffer::IndexBuffer(Device* device, std::vector<unsigned int> indices):
 	Buffer(device, sizeof(indices[0])*indices.size(), VK_BUFFER_USAGE_TRANSFER_DST_BIT | VK_BUFFER_USAGE_INDEX_BUFFER_BIT, VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT)
 {
